neteaseProject/requirements.txt
astroid==2.2.5
autopep8==1.4.3
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
Django==2.2.3
django-cors-headers==3.0.2
djangorestframework==3.10.2
idna==2.8
isort==4.3.16
kiwisolver==1.1.0
lazy-object-proxy==1.3.1
lxml==4.4.0
matplotlib==3.1.1
mccabe==0.6.1
numpy==1.17.0
pandas==0.25.0
Pillow==6.1.0
pycodestyle==2.5.0
pylint==2.3.1
PyMySQL==0.9.3
pyparsing==2.4.1.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
selenium==3.141.0
six==1.12.0
snownlp==0.12.3
soupsieve==1.9.2
sqlparse==0.3.0
typed-ast==1.3.1
urllib3==1.25.3
wrapt==1.11.1
xlrd==1.2.0
XlsxWriter==1.1.8
